+++ test/bract/core/inducer_test.clj
@@ -183,3 +183,11 @@
   (vreset! volatile-holder 0)
   (inducer/deinit {:bract.core/deinit (fn [] (vreset! volatile-holder 10))})
   (is (= 10 @volatile-holder)))
+
+
+(deftest test-invoke-stopper
+  (vreset! volatile-holder 0)
+  (inducer/invoke-stopper {})
+  (is (zero? @volatile-holder))
+  (inducer/invoke-stopper {:bract.core/stopper (fn [] (vswap! volatile-holder (fn [^long x] (inc x))))})
+  (is (= 1 @volatile-holder)))